==> causalPath_ananlysis_07-08-2021/1-data_preparation.Rmd <==
---
title: 'Preparing Trametinib and Gilteritinib Data'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
```

# Load data
```{r}
# Trametinib
tram <- read.csv("data/syn24389738.tsv", sep='\t')

# Gilteritinib
gilt <- read.csv("data/syn24189487.tsv", sep='\t', row.names = NULL)
```

# Create directory
```{r}
dir.create(path = "results/prepared", showWarnings = FALSE, recursive = TRUE)
```

# Prepare Trametinib Data
```{r}
# Remove unnecessary columns
tram_filt <- subset(tram, select = -Protein)

# Rename columns
colnames(tram_filt) <- c("gene", "site", "peptide", "logRatio", "sample", "cellLine", "timePoint", "treatment")

# Filter out MCP-1 treatments and Trametinib resistant cell lines
tram_filt <- tram_filt %>% 
  filter(treatment == c("none", "Trametinib")) %>% 
  filter(cellLine == "MOLM-13") %>% 
  arrange(gene, site)

# Replace auto-date correction
tram_filt$gene[tram_filt$gene == "2-Sep"] <- "SEPT2"
tram_filt$gene[tram_filt$gene == "6-Sep"] <- "SEPT6"
tram_filt$gene[tram_filt$gene == "7-Sep"] <- "SEPT7"

# Remove rows with empty gene values
tram_filt <- tram_filt[tram_filt$gene != "",]

# Set Na's to NaNs
tram_filt[is.na(tram_filt)] <- NaN

# Save
write.table(tram_filt, file = "results/prepared/trametinib.tsv", sep = "\t", row.names = FALSE)
```


# Prepare Gilteritinib Data
```{r}
# Trim unnecessary columns
gilt_filt <- subset(gilt, select = -c(ROW_ID, ROW_VERSION, Entry, ids))

# Rename columns
colnames(gilt_filt) <- c("gene", "site", "peptide", "sample", "logRatio", "cellLine", "ligand", "treatment", "timePoint")

# Replace DMSO with none
gilt_filt$treatment[gilt_filt$treatment == "DMSO"] <- "none"

# Filter out ligand-treated cells
gilt_filt_no_ligand <- gilt_filt %>% 
  filter(ligand == "None") %>% 
  arrange(gene, site)

# Remove ligand column
gilt_filt_no_ligand <- subset(gilt_filt_no_ligand, select = -ligand)

# Remove rows with empty gene values
gilt_filt_no_ligand <- gilt_filt_no_ligand[gilt_filt_no_ligand$gene != "",]

# Set Na's to NaNs
gilt_filt_no_ligand[is.na(gilt_filt_no_ligand)] <- NaN

# Save
write.table(gilt_filt_no_ligand, file = "results/prepared/gilteritinib.tsv", sep = "\t", row.names = FALSE)

# Extract MOLM14-a samples
molm14_a <- gilt_filt_no_ligand %>% 
  filter(cellLine == "MOLM14-a") 

# Save
write.table(molm14_a, file = "results/prepared/gilteritinib_molm14-a.tsv", sep = "\t", row.names = FALSE)

# Extract MOLM14-b samples
molm14_b <- gilt_filt_no_ligand %>% 
  filter(cellLine == "MOLM14-b") 

# Save
write.table(molm14_b, file = "results/prepared/gilteritinib_molm14-b.tsv", sep = "\t", row.names = FALSE)
```

# Session Information
```{r}
sessionInfo()
```
