==> causalPath_ananlysis_07-08-2021/2-sample_metadata.Rmd <==
---
title: 'Create metadata files'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
```

# Make directory
```{r}
# Make results directory
dir.create(path = "results/metadata", showWarnings = FALSE, recursive = TRUE)
```

# Tables to analyze
```{r}
expt_list <- c("trametinib", "gilteritinib", "gilteritinib_molm14-a", "gilteritinib_molm14-b")
```

# Create Metadata File
```{r}

for (expt in expt_list) {
  
  print(sprintf("Creating metadata table for %s", expt))
  
  # Load data
  data <- read.csv(sprintf("results/prepared/%s.tsv", expt), sep='\t')
  
  # Identify unique samples
  samples <- unique(data$sample)
  
  #Initialize metadata file
  metadata <- data.frame(matrix(nrow = length(samples), ncol = 4))
  colnames(metadata) <- c("sample", "timePoint", "treatment", "cellLine")
  metadata$sample <- samples
  
  # Extract metadata values for each sample
  for (sample in samples) {
    # Extract values
    timePoint <- data[data$sample == sample, "timePoint"][1]
    treatment <- data[data$sample == sample, "treatment"][1]
    cellLine <- data[data$sample == sample, "cellLine"][1]
    
    # Save in metadata file
    metadata[metadata$sample == sample, "timePoint"] <- timePoint
    metadata[metadata$sample == sample, "treatment"] <- treatment
    metadata[metadata$sample == sample, "cellLine"] <- cellLine
  }
  
  # Sort
  metadata <- metadata %>% 
    arrange(timePoint, treatment, cellLine)
  
  # Save
  write.table(metadata, sprintf("results/metadata/%s_metadata.tsv", expt), sep = "\t", row.names = FALSE)
}
```

# Session Information

```{r}
sessionInfo()
```
