==> causalPath_ananlysis_07-08-2021/3-cp_format.Rmd <==
---
title: 'Convert Data into the CausalPath Format'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
```

# Make directory
```{r}
# Make results directory
dir.create(path = "results/cp_format", showWarnings = FALSE, recursive = TRUE)
```

# Define functions

```{r}
# Create a function that adds peptide to the site to make a unique ID
# This will become the ID column
unique_id_fn <- function(x) {

  # Extract site and gene
  site <- data[x, "site"]
  gene <- data[x, "gene"]

  # Identify what will be the reference for the ID
  # The reference will be the site, unless a site isn't present
  if (site != "") {
    ref <- site
  } else {
    ref <- gene}

  # Extract and format peptide
  peptide <- data[x, "peptide"]
  if(peptide == NaN) {
    peptide <- ""
  }

  # Concatenate and return unique ID
  new <- paste0(ref, peptide, collapse="-")
  return(new)
}

# Create a function to remove lower case letters (s, t, y) and replace with hyphens
letter_replace_fn <- function(x) {
  if (x != "") {
      new <- str_split(x, "[a-z]")[[1]]
      new <- paste(new, collapse="-")
      new <- str_replace(new, pattern = "-$", replacement = "")
      return(new)
  } else {
    return(x)
  }
}

# Create a function to remove gene names from sites
# Replace hyphens with pipe symbols
# This will become the Site column
remove_genes_add_pipes_fn <- function(x) {
  
  # Extract site and gene
  site <- data[x, "site_letters_replaced"]
  gene <- data[x, "gene"]
  
  if (site != "") {
    # Remove gene
    no_genes <- str_split(site, paste0(gene, "-"))[[1]][2]
    
    # Replace hypthens with pipes
    no_genes_with_pipes <- str_replace_all(no_genes, pattern = "-", replacement = "|")
    
    return(no_genes_with_pipes)
  } else {
    
    return(site)
    
  }
  
}

```

# Tables to be re-formatted
```{r}
expt_list <- c("trametinib", "gilteritinib", "gilteritinib_molm14-a", "gilteritinib_molm14-b")
```

# Re-format each table

```{r}

for (expt in expt_list) {
  
  print(sprintf("Re-formatting the %s experiment into the causal path format", expt))
  
  # Load data
  data <- read.csv(sprintf("results/prepared/%s.tsv", expt), sep='\t')
  
  # Remove duplicates
  data <- data %>% distinct()
  
  # Make unique IDs
  print("Making unique IDs")
  rows <- seq(1:nrow(data))
  unique_id <- lapply(rows, unique_id_fn)
  data$unique_id <- unlist(unique_id)
  
  # Remove lower case letters (s, t, y) and replace with hyphen
  print("Removing lower case letters from sites")
  site_letters_replaced <- lapply(data$site, letter_replace_fn)
  data$site_letters_replaced <- unlist(site_letters_replaced)
  
  # Remove gene names from sites
  print("Removing gene names from sites")
  site_no_genes_with_pipes <- lapply(rows, remove_genes_add_pipes_fn)
  data$site_no_genes_with_pipes <- unlist(site_no_genes_with_pipes)
  
  # Add an Effect column
  data$Effect <- ""
  
  # Rename and rearrange columns
  prep_cp_format <- data %>% 
    rename(ID = unique_id,
           Sites = site_no_genes_with_pipes,
           Symbols = gene) %>% 
    select(ID, Sites, Effect, Symbols, sample, logRatio)
  
  # Replace NaN and NA values
  prep_cp_format$ID[( (is.nan(prep_cp_format$ID)) | (is.na(prep_cp_format$ID)) )] <- ""
  prep_cp_format$Sites[( (is.nan(prep_cp_format$Sites)) | (is.na(prep_cp_format$Sites)) )] <- ""
  prep_cp_format$Symbols[( (is.nan(prep_cp_format$Symbols)) | (is.na(prep_cp_format$Symbols)) )] <- ""
  
  # Remove duplicated rows
  prep_cp_format <- prep_cp_format %>% distinct()
  
  # Pivot wider
  print("Pivoting the table wider")
  cp_format <- pivot_wider(prep_cp_format, 
                           id_cols = c(ID, Sites, Effect, Symbols),
                           names_from = sample, 
                           values_from = logRatio)
  
  # Replace empty values
  cp_format[is.na(cp_format)] <- NaN
  
  # Save as a data.frame
  print("Saving")
  cp_format <- as.data.frame(cp_format)
  write.table(cp_format, file = sprintf("results/cp_format/%s_unnormalized.tsv", expt), sep = "\t", row.names = FALSE, quote = FALSE)
}
```

# Session Information

```{r}
sessionInfo()
```
