==> causalPath_ananlysis_07-08-2021/4-normalize.Rmd <==
---
title: 'Normalize Data (in CausalPath Format)'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
library(BBmisc)
library(factoextra)
```

# Create directory
```{r}
# Make results directory
dir.create(path = "results/normalize", showWarnings = FALSE, recursive = TRUE)
```

# Trametinib

## Load data
```{r}
# Drug condition
drug <- "trametinib"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
data_norm <- normalize(data)
```

## Average across replicates
```{r}
# Calculate row means excluding NA's
time0 <- rowMeans(data_norm[, c("X0_P1_01", "X0_P3_24", "X0_P2_20")], na.rm = TRUE)
time5 <- rowMeans(data_norm[, c("X5T_P1_04", "X5T_P3_25", "X5T_P2_12")], na.rm = TRUE)
time60 <- rowMeans(data_norm[, c("X60T_P1_05", "X60T_P3_28", "X60T_P2_13")], na.rm = TRUE)

# Column bind new columns
norm_vals <- as.data.frame(cbind(unlist(time0), unlist(time5), unlist(time60)))
colnames(norm_vals) <- c("time0", "time5", "time60")
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.nan(norm_vals$time0) & is.nan(norm_vals$time5) & is.nan(norm_vals$time60) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.nan(norm_vals$time0) | is.nan(norm_vals$time5) | is.nan(norm_vals$time60) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm_avg <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_avg_filt <- data_norm_avg[ !(rownames(data_norm_avg) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_avg_filt[is.na(data_norm_avg_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_avg_filt[,5:ncol(data_norm_avg_filt)]
samples <- colnames(data_norm_avg_filt[,5:ncol(data_norm_avg_filt)])
groups <- as.factor(c(rep("dmso", 1), rep("trametinib", 2)))
features <- data_norm_avg_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_normalized_pca.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_avg_filt, file = sprintf("results/normalize/%s_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Trametinib (all reps)

## Load data
```{r}
# Drug condition
drug <- "trametinib"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.na(norm_vals$X0_P1_01) & is.na(norm_vals$X0_P3_24) & is.na(norm_vals$X5T_P1_04) & is.na(norm_vals$X5T_P3_25) & is.na(norm_vals$X60T_P1_05) & is.na(norm_vals$X60T_P3_28) & is.na(norm_vals$X0_P2_20) & is.na(norm_vals$X5T_P2_12) & is.na(norm_vals$X60T_P2_13) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
nPeptides <- nrow(is.na(norm_vals))
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
groups <- as.factor(c("none", "none", "trametinib-5", "trametinib-5", "trametinib-60", "trametinib-60", "none", "trametinib-5", "trametinib-60"))
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07", "#800080"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_all_reps_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib (all)

## Load data
```{r}
# Drug condition
drug <- "gilteritinib"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
data_norm <- normalize(data)
```

## Average across replicates
```{r}
# Calculate row means excluding NA's
dmso_0 <- rowMeans(data_norm[, c("P1", "P6")], na.rm = TRUE)
dmso_30 <- rowMeans(data_norm[, c("P2", "P7")], na.rm = TRUE)
dmso_180 <- rowMeans(data_norm[, c("P4", "P9")], na.rm = TRUE)
gilt_30 <- rowMeans(data_norm[, c("P3", "P8")], na.rm = TRUE)
gilt_180 <- rowMeans(data_norm[, c("P5", "P10")], na.rm = TRUE)

# Column bind new columns
norm_vals <- as.data.frame(cbind(unlist(dmso_0), unlist(dmso_30), unlist(dmso_180), unlist(gilt_30), unlist(gilt_180)))
colnames(norm_vals) <- c("dmso_0", "dmso_30", "dmso_180", "gilt_30", "gilt_180")
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.nan(norm_vals$dmso_0) & is.nan(norm_vals$dmso_30) & is.nan(norm_vals$dmso_180) & is.nan(norm_vals$gilt_30) & is.nan(norm_vals$gilt_180) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.nan(norm_vals$dmso_0) | is.nan(norm_vals$dmso_30) | is.nan(norm_vals$dmso_180) | is.nan(norm_vals$gilt_30) | is.nan(norm_vals$gilt_180) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm_avg <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_avg_filt <- data_norm_avg[ !(rownames(data_norm_avg) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_avg_filt[is.na(data_norm_avg_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_avg_filt[,5:ncol(data_norm_avg_filt)]
samples <- colnames(data_norm_avg_filt[,5:ncol(data_norm_avg_filt)])
features <- data_norm_avg_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
# By drug condition
groups <- as.factor(c(rep("dmso", 3), rep("gilteritinib", 2)))
fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_normalized_pca.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_avg_filt, file = sprintf("results/normalize/%s_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib MOLM14-a

## Load data
```{r}
# Drug condition
drug <- "gilteritinib_molm14-a"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.nan(norm_vals$P1) & is.nan(norm_vals$P2) & is.nan(norm_vals$P3) & is.nan(norm_vals$P4) & is.nan(norm_vals$P5) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.nan(norm_vals$P1) | is.nan(norm_vals$P2) | is.nan(norm_vals$P3) | is.nan(norm_vals$P4) | is.nan(norm_vals$P5) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
groups <- as.factor(c("dmso", "dmso", "gilteritinib", "dmso", "gilteritinib"))
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_normalized_pca.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib MOLM14-b

## Load data
```{r}
# Drug condition
drug <- "gilteritinib_molm14-b"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.nan(norm_vals$P10) & is.nan(norm_vals$P6) & is.nan(norm_vals$P7) & is.nan(norm_vals$P8) & is.nan(norm_vals$P9) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.nan(norm_vals$P10) | is.nan(norm_vals$P6) | is.nan(norm_vals$P7) | is.nan(norm_vals$P8) | is.nan(norm_vals$P9) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
groups <- as.factor(c("gilteritinib", "dmso", "dmso", "gilteritinib", "dmso"))
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_normalized_pca.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib (all reps)

## Load data
```{r}
# Drug condition
drug <- "gilteritinib"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.nan(norm_vals$P10) & is.nan(norm_vals$P6) & is.nan(norm_vals$P7) & is.nan(norm_vals$P8) & is.nan(norm_vals$P9) & is.nan(norm_vals$P1) & is.nan(norm_vals$P2) & is.nan(norm_vals$P3) & is.nan(norm_vals$P4) & is.nan(norm_vals$P5) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.na(norm_vals) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
# By drug conditions
groups <- as.factor(c("none-0", "gilteritinib-180", "none-30", "gilteritinib-30", "none-180", "gilteritinib-180", "none-0", "none-30", "gilteritinib-30", "none-180"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07", "#800080", "#00FF00", "#00FFFF"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_conditions.png", drug), device = "png")

# Visualize PCA
# By cell lines
groups <- as.factor(c("MOLM14-a", "MOLM14-b", "MOLM14-a", "MOLM14-a", "MOLM14-a", "MOLM14-a", "MOLM14-b", "MOLM14-b", "MOLM14-b", "MOLM14-b"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_cell_lines.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_all_reps_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib MOLM-14a (all reps)

## Load data
```{r}
# Drug condition
drug <- "gilteritinib_molm14-a"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.na(norm_vals$P1) & is.na(norm_vals$P2) & is.na(norm_vals$P3) & is.na(norm_vals$P4) & is.na(norm_vals$P5) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.na(norm_vals) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
# By drug conditions
groups <- as.factor(c("none-0", "none-30", "gilteritinib-30", "none-180", "gilteritinib-180"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07", "#800080", "#00FF00", "#00FFFF"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_conditions.png", drug), device = "png")

# Visualize PCA
# By cell lines
groups <- as.factor(c("none", "none", "gilteritinib", "none", "gilteritinib"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_cell_lines.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_all_reps_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Gilteritinib MOLM-14b (all reps)

## Load data
```{r}
# Drug condition
drug <- "gilteritinib_molm14-b"

# Load data
data <- read.csv(sprintf("results/cp_format/%s_unnormalized.tsv", drug) , sep='\t')
```

## Column normalize
```{r}
# Only normalize sample values
norm_vals <- normalize(data[,5:ncol(data)])
```

## Address Sparsity
```{r}
# Identify the number of peptides still have NaN's
null_vals <- norm_vals[ is.na(norm_vals$P1) & is.na(norm_vals$P2) & is.na(norm_vals$P3) & is.na(norm_vals$P4) & is.na(norm_vals$P5) ,]
null_idx <- rownames(null_vals)
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that have all NaN values", nPeptides, round(pctTotal * 100,2), nTotal))

# Identify the number of peptides that have at least one NaN
null_vals <- norm_vals[ is.na(norm_vals) ,]
nPeptides <- nrow(null_vals)
nTotal <- nrow(norm_vals)
pctTotal <- nPeptides / nTotal
print(sprintf("There are %s peptides (%s%% of %s total) that at least one NaN value", nPeptides, round(pctTotal * 100,2), nTotal))

# Make sure NA values in first four columns are blank
first_cols <- data[,1:4]
first_cols[is.na(first_cols)] <- ""

# Combine new columns with causal path columns
data_norm <- cbind(first_cols, norm_vals)

# Remove peptides with all NaN's
data_norm_filt <- data_norm[ !(rownames(data_norm) %in% null_idx) ,]

# Replace remaining NaN's with zero
data_norm_filt[is.na(data_norm_filt)] <- 0
```

## Principle Component Analysis (PCA)
```{r}
# Extract data for PCA analysis
pca_data <- data_norm_filt[,5:ncol(data_norm_filt)]
samples <- colnames(data_norm_filt[,5:ncol(data_norm_filt)])
features <- data_norm_filt$ID

# Transpose and save as data frame
pca_data <- as.data.frame(t(pca_data))
rownames(pca_data) <- samples
colnames(pca_data) <- features

# Calculate PCA
pca <- prcomp(pca_data, scale.=TRUE, center = TRUE)

# Percent explained variance
fviz_eig(pca)
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_variance.png", drug), device = "png")

# Visualize PCA
# By drug conditions
groups <- as.factor(c("gilteritinib-180", "none-0", "none-30", "gilteritinib-30", "none-180"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07", "#800080", "#00FF00", "#00FFFF"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_conditions.png", drug), device = "png")

# Visualize PCA
# By cell lines
groups <- as.factor(c("none", "none", "gilteritinib", "none", "gilteritinib"))

fviz_pca_ind(pca,
             col.ind = groups, 
             palette = c("#00AFBB",  "#FC4E07"),
             legend.title = "Groups",
             repel = TRUE
             )
ggsave(sprintf("results/normalize/%s_all_reps_normalized_pca_cell_lines.png", drug), device = "png")
```

## Save normalized data
```{r}
# Save
write.table(data_norm_filt, file = sprintf("results/normalize/%s_all_reps_normalized.tsv", drug), sep = "\t", row.names = FALSE, quote = FALSE)
```

# Session Information

```{r}
sessionInfo()
```
