==> causalPath_ananlysis_07-08-2021/5-parameter_file.Rmd <==
---
title: 'Create Parameter Files and Folders for CausalPath Analysis'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
```

# Parameter file function
```{r}
parameter_files <- function(pair_list) {
  
  for (pair in pair_list) {
    
    print(sprintf("Creating parameter files for the %s and %s samples treated with %s", pair[[1]], pair[[2]], group))
  
    # create directory
    dir_path <- sprintf("results/causalpath/%s/%s", group, paste(pair[[1]], pair[[2]], sep = "-"))
    dir.create(path = dir_path, showWarnings = FALSE, recursive = TRUE)
    
    # Define the file used for causalpath analysis
    cp_file <- sprintf("%s_normalized.tsv", group)
    
    # Set parameters
    # parameters = c(paste0("proteomics-values-file = ", cp_file),
    #                "id-column = ID",
    #                "symbols-column = Symbols",
    #                "sites-column = Sites",
    #                "effect-column = Effect",
    #                "value-transformation = significant-change-of-mean",
    #                "fdr-threshold-for-data-significance = 0.1 phosphoprotein",
    #                "fdr-threshold-for-data-significance = 0.1 protein",
    #                "calculate-network-significance = true",
    #                "permutations-for-significance = 100",
    #                "generate-data-centric-graph = false",
    #                "show-insignificant-data = false",
    #                "color-saturation-value = 50",
    #                "custom-resource-directory = C:/Users/wmyas/gitPackages/causalpath/.panda",
    #                "use-network-significance-for-causal-reasoning = false")
    
    parameters = c(paste0("proteomics-values-file = ", cp_file),
               "id-column = ID",
               "symbols-column = Symbols",
               "sites-column = Sites",
               "effect-column = Effect",
               "value-transformation = fold-change-of-mean",
               "threshold-for-data-significance = 2.5 phosphoprotein",
               "threshold-for-data-significance = 2.5 protein",
               "calculate-network-significance = true",
               "permutations-for-significance = 100",
               "generate-data-centric-graph = false",
               "show-insignificant-data = false",
               "color-saturation-value = 50",
               "custom-resource-directory = C:/Users/wmyas/gitPackages/causalpath/.panda",
               "use-network-significance-for-causal-reasoning = false")
    
    # Set control values
    parameters[ length(parameters) + 1 ] <- paste0("control-value-column = ", pair[[1]])
    
    # Set test columns
    parameters[ length(parameters) + 1 ] <- paste0("test-value-column = ", pair[[2]])
    
    # Write to parameters file
    file_path <- file(sprintf("%s/parameters.txt", dir_path))
    writeLines(parameters, file_path)
    close(file_path)
    
    # Save Causal Path File
    write.table(data, file = sprintf("%s/%s", dir_path, cp_file), sep = "\t", row.names = FALSE, quote = FALSE)
    
  }
  
}

parameter_files_multiple_conds <- function(controls, tests, time_point) {
    
  print(sprintf("Creating parameter files for the %s samples at %s minutes", group, time_point))

  # create directory
  dir_path <- sprintf("results/causalpath/%s/%s", group, paste("control", sprintf("test%s", time_point), sep = "-"))
  dir.create(path = dir_path, showWarnings = FALSE, recursive = TRUE)
  
  # Define the file used for causalpath analysis
  cp_file <- sprintf("%s_normalized.tsv", group)
  
  # Set parameters
  parameters = c(paste0("proteomics-values-file = ", cp_file),
                 "id-column = ID",
                 "symbols-column = Symbols",
                 "sites-column = Sites",
                 "effect-column = Effect",
                 "value-transformation = significant-change-of-mean",
                 "fdr-threshold-for-data-significance = 0.2 phosphoprotein",
                 "fdr-threshold-for-data-significance = 0.2 protein",
                 "calculate-network-significance = true",
                 "permutations-for-significance = 100",
                 "generate-data-centric-graph = false",
                 "show-insignificant-data = false",
                 "color-saturation-value = 50",
                 "custom-resource-directory = C:/Users/wmyas/gitPackages/causalpath/.panda",
                 "use-network-significance-for-causal-reasoning = false")
  
  # parameters = c(paste0("proteomics-values-file = ", cp_file),
  #            "id-column = ID",
  #            "symbols-column = Symbols",
  #            "sites-column = Sites",
  #            "effect-column = Effect",
  #            "value-transformation = fold-change-of-mean",
  #            "threshold-for-data-significance = 2 phosphoprotein",
  #            "threshold-for-data-significance = 2 protein",
  #            "calculate-network-significance = true",
  #            "permutations-for-significance = 100",
  #            "generate-data-centric-graph = false",
  #            "show-insignificant-data = false",
  #            "color-saturation-value = 50",
  #            "custom-resource-directory = C:/Users/wmyas/gitPackages/causalpath/.panda",
  #            "use-network-significance-for-causal-reasoning = false")
  
  # Set control values
  for (control in controls) {
    parameters[ length(parameters) + 1 ] <- paste0("control-value-column = ", control) 
  }
  
  # Set test columns
  for (test in tests) {
    parameters[ length(parameters) + 1 ] <- paste0("test-value-column = ", test) 
  }
  
  # Write to parameters file
  file_path <- file(sprintf("%s/parameters.txt", dir_path))
  writeLines(parameters, file_path)
  close(file_path)
  
  # Save Causal Path File
  write.table(data, file = sprintf("%s/%s", dir_path, cp_file), sep = "\t", row.names = FALSE, quote = FALSE)
  
}
```

# Trametinib

## Load data
```{r}
group <- "trametinib"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)
data$Effect <- ""
```

## Parameter files
```{r}
# Pairwise comparisons
# List the names of the control and test columns, respectively
pair1 <- list("time0", "time5") # DMSO time 0 vs Trametinib time 5 min
pair2 <- list("time0", "time60") # DMSO time 0 vs Trametinib time 60 min
pair_list <- list(pair1, pair2)

# Call parameter file function
parameter_files(pair_list)
```

# Gilteritinib (all)

## Load data
```{r}
group <- "gilteritinib"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)
data$Effect <- ""
```

## Parameter files
```{r}
# Pairwise comparisons
# List the names of the control and test columns, respectively
pair1 <- list("dmso_0", "gilt_30") # DMSO time 0 vs Gilteritinib time 30 min
pair2 <- list("dmso_30", "gilt_30") # DMSO time 30 vs Gilteritinib time 30 min
pair3 <- list("dmso_0", "gilt_180") # DMSO time 0 vs Gilteritinib time 180 min
pair4 <- list("dmso_180", "gilt_180") # DMSO time 180 vs Gilteritinib time 180 min
pair_list <- list(pair1, pair2, pair3, pair4)

# Call parameter file function
parameter_files(pair_list)
```

## Parameter files
```{r}
# Control samples
controls <- list("dmso_0", "dmso_30", "dmso_180")

# Test samples
gilt_30 <- list("gilt_30")
gilt_180 <- list("gilt_180")

# Call parameter file function
parameter_files_multiple_conds(controls, gilt_30, 30)
parameter_files_multiple_conds(controls, gilt_180, 180)
```

# Gilteritinib - MOLM-14a

## Load data
```{r}
group <- "gilteritinib_molm14-a"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)

# Ensure empty space in Sites and Effects column are properly formatted
data$Effect <- ""
data[is.na(data), "Sites"] <- ""

# Ensure remaining NA values are NaNs
data[is.na(data), ] <- NaN
```

## Parameter files
```{r}
# Pairwise comparisons
# List the names of the control and test columns, respectively
pair1 <- list("P1", "P3") # DMSO time 0 vs Gilteritinib time 30 min
pair2 <- list("P2", "P3") # DMSO time 30 min vs Gilteritinib time 30 min
pair3 <- list("P1", "P5") # DMSO time 0 vs Gilteritinib time 180 min
pair4 <- list("P4", "P5") # DMSO time 180 vs Gilteritinib time 180 min

pair_list <- list(pair1, pair2, pair3, pair4)

# Call parameter file function
parameter_files(pair_list)
```

# Gilteritinib - MOLM-14b

## Load data
```{r}
group <- "gilteritinib_molm14-b"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)

# Ensure empty space in Sites and Effects column are properly formatted
data$Effect <- ""
data[is.na(data), "Sites"] <- ""

# Ensure remaining NA values are NaNs
data[is.na(data), ] <- NaN
```

## Parameter files
```{r}
# Pairwise comparisons
# List the names of the control and test columns, respectively
pair1 <- list("P6", "P8") # DMSO time 0 vs Gilteritinib time 30 min
pair2 <- list("P7", "P8") # DMSO time 30 min vs Gilteritinib time 30 min
pair3 <- list("P6", "P10") # DMSO time 0 vs Gilteritinib time 180 min
pair4 <- list("P9", "P10") # DMSO time 180 vs Gilteritinib time 180 min

pair_list <- list(pair1, pair2, pair3, pair4)

# Call parameter file function
parameter_files(pair_list)
```

# Gilteritinib (all reps)

## Load data
```{r}
group <- "gilteritinib_all_reps"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)

# Ensure empty space in Sites and Effects column are properly formatted
data$Effect <- ""
data[is.na(data), "Sites"] <- ""

# Ensure remaining NA values are NaNs
data[is.na(data), ] <- NaN
```

## Parameter files
```{r}
# Control samples
controls <- list("P1", "P2", "P4", "P6", "P7", "P9")

# Test samples
gilt_30 <- list("P3", "P8")
gilt_180 <- list("P5", "P10")

# Call parameter file function
parameter_files_multiple_conds(controls, gilt_30, 30)
parameter_files_multiple_conds(controls, gilt_180, 180)
```

# Gilteritinib MOLM14-a (all reps)

## Load data
```{r}
group <- "gilteritinib_molm14-a_all_reps"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)

# Ensure empty space in Sites and Effects column are properly formatted
data$Effect <- ""
data[is.na(data), "Sites"] <- ""

# Ensure remaining NA values are NaNs
data[is.na(data), ] <- NaN
```

## Parameter files
```{r}
# Control samples
controls <- list("P1", "P2", "P4")

# Test samples
gilt_30 <- list("P3")
gilt_180 <- list("P5")

# Call parameter file function
parameter_files_multiple_conds(controls, gilt_30, 30)
parameter_files_multiple_conds(controls, gilt_180, 180)
```

# Gilteritinib MOLM14-b (all reps)

## Load data
```{r}
group <- "gilteritinib_molm14-b_all_reps"

# Load data
# Ensure Effect is empty whitespace
data <- read.table(file = sprintf("results/normalize/%s_normalized.tsv", group), sep = "\t", header = TRUE, check.names = FALSE)

# Ensure empty space in Sites and Effects column are properly formatted
data$Effect <- ""
data[is.na(data), "Sites"] <- ""

# Ensure remaining NA values are NaNs
data[is.na(data), ] <- NaN
```

## Parameter files
```{r}
# Control samples
controls <- list("P6", "P7", "P9")

# Test samples
gilt_30 <- list("P8")
gilt_180 <- list("P10")

# Call parameter file function
parameter_files_multiple_conds(controls, gilt_30, 30)
parameter_files_multiple_conds(controls, gilt_180, 180)
```

# Session Information

```{r}
sessionInfo()
```
