==> causalPath_ananlysis_07-08-2021/6-batch_correction.Rmd <==
---
title: 'Create Parameter Files and Folders for CausalPath Analysis'
author: "Will Yashar"
date: "`r Sys.Date()`"
output:
  html_document:
    code_folding: show
    number_sections: no
    theme: lumen
    toc: yes
    toc_float:
      collapsed: yes
editor_options:
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

# Packages
```{r}
library(tidyverse)
library(PhosR)
library(stringr)
```

# Load data
```{r}
data <- read.table(file = "results/cp_format/gilteritinib_unnormalized.tsv", sep = "\t", header = TRUE, check.names = FALSE)
data$Effect <- ""
```

# Make directory
```{r}
dir.create("results/batch_corrected", recursive = TRUE, showWarnings = FALSE)
```

# Format data
```{r}
# Create rows for each unique phosphorylation site
# Based on: https://pyanglab.github.io/PhosR/articles/PhosR.html 

# Separate phosphosites by pipe symbol
sep <- separate(data = data, col = Sites, into = c("first", "second", "third"), sep = "\\|")

# Extract sequences
sep$Sequences <- str_extract(data$ID, "\\S\\.\\S+")

# Two sites in same row indicates Mass Spec wasn't able to discriminate between sites
# Assume both have same expression
sep_long <- sep %>% 
  pivot_longer(cols = c("first", "second", "third"), values_to = "Sites") %>% 
  drop_na(Sites)

# Update ID's
for (i in 1:nrow(sep_long)) {
  
  symbol <- sep_long[i,"Symbols"]
  site <- sep_long[i, "Sites"]
  sequ <- sep_long[i, "Sequences"]
  
  if (!is.na(site)) {
    
    new <- paste(symbol, site, sequ, sep = "-")
    
  } else {
    
    new <- paste(symbol, sequ, sep = "-")
  }
  
  sep_long[i,"ID"] <- new
  
}

# Re-arrange columns
# Remove duplicate rows
sep_long <- sep_long %>% 
  select(-c(name,Sequences)) %>% 
  relocate(c("ID", "Sites", "Effect", "Symbols"), .before = P1) 
```

# Create PhosphoExperiment object
```{r}
# Extract quantification data
quant <- sep_long[,5:ncol(sep_long)]
rownames(quant) <- sep_long$ID

# Create PhosphoExperiment object
ppe <- PhosphoExperiment(assays = list(Quantification = as.matrix(quant)))

# Add annotation
GeneSymbol(ppe) <- sep_long$Symbols
Sequence(ppe) <- str_extract(sep_long$ID, "\\S\\.\\S+")
Site(ppe) <- as.numeric(gsub("[A-Z]","", sapply(sep_long$Sites, "[[", 1)))
Residue(ppe) <- sep_long$Sites # Storing the full site names in the residue slot as Sites has to be numeric

# Check number of peptides
dim(ppe)
```

# Filter 
```{r}
# Define groups
grps <- c("none", "gilteritinib-180", "none", "gilteritinib-30", "none",
          "gilteritinib-180", "none", "none", "gilteritinib-30", "none")

# Ensure phosphopeptides are in at least half of the replicates in at least one of the conditions
ppe_filt <- selectGrps(ppe, grps, 0.5, n=1)

# Check number of peptides
dim(ppe_filt)
```

# Imputation
```{r}
# Tail-based imputation
set.seed(314159)
ppe_imputed <- tImpute(ppe_filt)

# Save data
imputed_data <- as.data.frame(ppe_imputed@assays@data$imputed)
imputed_data <- imputed_data %>% 
  rownames_to_column("ID") %>% 
  mutate(Effect = "",
         Sites = ppe_imputed@Residue,
         Symbols = ppe_imputed@GeneSymbol) %>% 
  relocate(c("ID", "Sites", "Effect", "Symbols"), .before = P1)

write.table(imputed_data, "results/batch_corrected/gilteritinib_imputed.tsv", sep = "\t", row.names = FALSE, quote = FALSE)
```

# Quantification Plots
```{r}
# Percentage of quantified sites after filtering and imputation
p1 = plotQC(SummarizedExperiment::assay(ppe_filt,"Quantification"), 
    labels=colnames(ppe_filt), panel = "quantify", grps = grps) +
    ggplot2::ggtitle("Quantification per sample (pre-imputation)")

p2 = plotQC(SummarizedExperiment::assay(ppe_imputed,"imputed"), 
    labels=colnames(ppe_imputed), panel = "quantify", grps = grps) +
    ggplot2::ggtitle("Quantification per sample (post-imputation)")

ggpubr::ggarrange(p1, p2, nrow = 1)
ggsave("results/batch_corrected/gilteritinib_quantification_plots.png", dev = "png")
```

# PCA Pre-Scaling
```{r}
# By drug condition
plotQC(SummarizedExperiment::assay(ppe_imputed, "imputed"), grps=grps, 
    labels = colnames(ppe_imputed), panel = "pca") +
    ggplot2::ggtitle("Before scaling")
ggsave("results/batch_corrected/gilteritinib_pca_by_drug_pre_scaling.png", dev = "png")

# By cell line
cell_line_grps <- c("MOLM14-a", "MOLM14-b", "MOLM14-a", "MOLM14-a", "MOLM14-a", 
                    "MOLM14-a", "MOLM14-b", "MOLM14-b", "MOLM14-b", "MOLM14-b") 
plotQC(SummarizedExperiment::assay(ppe_imputed,"imputed"), grps=cell_line_grps, 
    labels = colnames(ppe_imputed), panel = "pca") +
    ggplot2::ggtitle("Before scaling")
ggsave("results/batch_corrected/gilteritinib_pca_by_cell_line_pre_scaling.png", dev = "png")
```

# Scaling
```{r}
# Median scaling (although workflow says normalization?)
ppe_imputed_scaled <- medianScaling(ppe_imputed, scale = TRUE, grps = grps, assay = "imputed")

# Save
scaled_data <- as.data.frame(ppe_imputed_scaled@assays@data$scaled)
scaled_data <- scaled_data %>% 
  rownames_to_column("ID") %>% 
  mutate(Effect = "",
         Sites = ppe_imputed_scaled@Residue,
         Symbols = ppe_imputed_scaled@GeneSymbol) %>% 
  relocate(c("ID", "Sites", "Effect", "Symbols"), .before = P1)

write.table(scaled_data, "results/batch_corrected/gilteritinib_imputed_scaled.tsv", sep = "\t", row.names = FALSE, quote = FALSE)
```

# PCA Post-Scaling Correction
```{r}
# By drug condition
plotQC(SummarizedExperiment::assay(ppe_imputed_scaled,"scaled"), grps=grps, 
    labels = colnames(ppe_imputed_scaled), panel = "pca") +
    ggplot2::ggtitle("After scaling and before batch correction")
ggsave("results/batch_corrected/gilteritinib_pca_by_drug_post_scaling.png", dev = "png")

# By cell line
cell_line_grps <- c("MOLM14-a", "MOLM14-b", "MOLM14-a", "MOLM14-a", "MOLM14-a", 
                    "MOLM14-a", "MOLM14-b", "MOLM14-b", "MOLM14-b", "MOLM14-b") 
plotQC(SummarizedExperiment::assay(ppe_imputed_scaled,"scaled"), grps=cell_line_grps, 
    labels = colnames(ppe_imputed_scaled), panel = "pca") +
    ggplot2::ggtitle("After scaling and before batch correction")
ggsave("results/batch_corrected/gilteritinib_pca_by_cell_line_post_scaling.png", dev = "png")
```

# Batch Correction & Normalization
```{r}
# Design matrix by cell line
design <- model.matrix( ~ cell_line_grps - 1)

# Stable phosphosites
# Use PhosR reference
data(SPSs)

# Extract sites
sites = paste(sapply(GeneSymbol(ppe_imputed_scaled), function(x)x),";",
              sapply(Residue(ppe_imputed_scaled), function(x)x), ";", sep = "")

ctl <- which( sites %in% SPSs)

# Batch correct
ppe_imputed_scaled_batch <- RUVphospho(ppe_imputed_scaled, M = design, k = 2, ctl = ctl)

# Save
normalized_data <- as.data.frame(ppe_imputed_scaled_batch@assays@data$normalised)
normalized_data <- normalized_data %>% 
  rownames_to_column("ID") %>% 
  mutate(Effect = "",
         Sites = ppe_imputed_scaled_batch@Residue,
         Symbols = ppe_imputed_scaled_batch@GeneSymbol) %>% 
  relocate(c("ID", "Sites", "Effect", "Symbols"), .before = P1)

write.table(scaled_data, "results/batch_corrected/gilteritinib_imputed_scaled_batch_corrected.tsv", sep = "\t", row.names = FALSE, quote = FALSE)
```

# PCA Post-Batch Correction
```{r}
# By drug condition
plotQC(SummarizedExperiment::assay(ppe_imputed_scaled_batch,"normalised"), grps=grps, 
    labels = colnames(ppe_imputed_scaled_batch), panel = "pca") +
    ggplot2::ggtitle("After batch correction")
ggsave("results/batch_corrected/gilteritinib_pca_by_drug_post_batch_correction.png", dev = "png")

# By cell line
cell_line_grps <- c("MOLM14-a", "MOLM14-b", "MOLM14-a", "MOLM14-a", "MOLM14-a", 
                    "MOLM14-a", "MOLM14-b", "MOLM14-b", "MOLM14-b", "MOLM14-b") 
plotQC(SummarizedExperiment::assay(ppe_imputed_scaled_batch,"normalised"), grps=cell_line_grps, 
    labels = colnames(ppe_imputed_scaled_batch), panel = "pca") +
    ggplot2::ggtitle("After batch correction")
ggsave("results/batch_corrected/gilteritinib_pca_by_cell_line_post_batch_correction.png", dev = "png")
```

# Hierarchical Clustering
```{r}
# Clustering after filtering and imputation
p1 = plotQC(SummarizedExperiment::assay(ppe_filt,"Quantification"), 
       labels=colnames(ppe_filt), panel = "dendrogram", grps = grps) +
    ggplot2::ggtitle("Pre-imputation")

p2 = plotQC(SummarizedExperiment::assay(ppe_imputed_scaled,"scaled"), 
       labels=colnames(ppe_imputed_scaled), panel = "dendrogram", grps = grps)+
    ggplot2::ggtitle("Post-scaling")

p3 = plotQC(SummarizedExperiment::assay(ppe_imputed_scaled_batch,"normalised"), 
       labels=colnames(ppe_imputed_scaled_batch), panel = "dendrogram", grps = grps)+
    ggplot2::ggtitle("Post-batch correction")

ggpubr::ggarrange(p1, p2, p3, nrow = 1)
ggsave("results/batch_corrected/gilteritinib_clustering.png", dev = "png")
```
